requirements.txt
asgiref==3.6.0
cached-property==1.5.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudinary==1.32.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
dj-config-url==0.1.1
Django==4.2.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.0.0
django-cpf-cnpj==1.0.0
django-crispy-forms==2.0
django-discover-runner==1.0
django-filter==23.2
django-role-permissions==3.1.1
django-url-filter==0.3.15
djangorestframework==3.14.0
drf-yasg==1.21.5
enum-compat==0.0.3
Flask==2.2.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.1.0
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
nodeenv==1.7.0
npm==0.1.1
numpy==1.24.3
opencv-contrib-python-headless==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.0
Pillow==9.5.0
prettyconf==2.2.1
psycopg2-binary==2.9.6
pybind11==2.10.4
pytz==2023.2
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
shapely==2.0.1
six==1.16.0
skia-python==87.5
sqlparse==0.4.3
tzdata==2023.2
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.2.3
xlrd==2.0.1
zipp==3.15.0
gevent
psycogreen
django-db-geventpool
Brotli==1.0.9
crispy-bootstrap4==2022.1
crispy-forms-materialize==0.2
django-materialize==1.0.1a2
greenlet==2.0.2
zope.event==4.6
zope.interface==6.0
sentry-sdk==1.22.2
celery==5.2.7
whitenoise==6.4.0
boto3==1.26.133
django-storages==1.13.2